backend/requirements.txt
PyPDF2
pdfplumber
PyMuPDF
litellm
langchain
langchain-core
langgraph
fastAPI
uvicorn
python-multipart
google-cloud-storage
python-dotenv
google-cloud-secret-manager>=2.16.0